config slice renamed to global
